// src/shared/enums/application-status.ts
export enum ApplicationStatus {
  BOOTSTRAPPING = 'BOOTSTRAPPING',
  PREFLIGHT_CHECKING = 'PREFLIGHT_CHECKING',
  STEAMCMD_MISSING = 'STEAMCMD_MISSING',
  STEAMCMD_INSTALLING = 'STEAMCMD_INSTALLING',
  STEAMCMD_READY = 'STEAMCMD_READY',
  SERVER_MISSING = 'SERVER_MISSING',
  SERVER_INSTALLING = 'SERVER_INSTALLING',
  SERVER_READY = 'SERVER_READY',
  CONFIGURATION_MISSING = 'CONFIGURATION_MISSING',
  CONFIGURATION_READY = 'CONFIGURATION_READY',
  FIREWALL_MISSING = 'FIREWALL_MISSING',
  READY = 'READY',
  SERVER_STARTING = 'SERVER_STARTING',
  SERVER_RUNNING = 'SERVER_RUNNING',
  SERVER_STOPPING = 'SERVER_STOPPING',
  ERROR = 'ERROR'
}